Revamp the design of the calculator
By utilising grid, I added background elements that would copy what was
in the display at any time. Had to create a new function and replace
some lines that involved calc_display.textContent. Also, increased the
string length for the calc_display from 9 to 12, and the exponential
notation from 4 to 6, to suit the new design.